Debugging a full run-through of translating the handbook. This has lead to:  - adding preservation symbols  - setting the max field length in the csv read operations (check later if this must be kept)  - corrections to the swapping function
Committing here before restructuring the translation dictionary, which is needed to handle the following issue:
When the same paragraph appears more than once in the document and has a different run anatomy,
then the translation dictionary will only have a single key-value pair to handle those paragraphs.
The problem is that each paragraph should have unique handling, since the run anatomy is unique.
The current setup leads to index out-of-range exceptions when run tags refer to a run which was
present for one paragraph but not another.